requirements.txt
mem0ai>=0.1.0
flask>=2.0.0
requests>=2.25.0
httpx>=0.21.0
